changed lastFrameTime in FrameLimiter to double, used mbstowcs_s() instead of mbstowcs() in create_wcharstr function

+++ D3DEngine/Source/framelimiter.cpp
@@ -60,7 +60,7 @@ void FrameLimiter::end()
 	m_buffers->lastFrameTime = elapsed_milli.count();
 }
 
-float FrameLimiter::getLastFramTime()
+double FrameLimiter::getLastFramTime()
 {
 	return m_buffers->lastFrameTime / 1000.0;
 }
